requirements.txt
PyQt6>=6.6.0
mss>=9.0.1
numpy>=1.24.0
Pillow>=10.0.0
pyaudio>=0.2.14
sounddevice>=0.4.6
pywin32>=306

